+++ test/utils/test_to_dense_batch.py
@@ -54,34 +54,26 @@ def test_to_dense_batch(fill):
     out, mask = to_dense_batch(x, batch, batch_size=4, fill_value=fill)
     assert out.size() == (4, 3, 2)
 
-    with set_experimental_mode(True, "disable_dynamic_shapes"):
-        with pytest.raises(ValueError):
-            out, mask = to_dense_batch(x, batch, max_num_nodes=6,
-                                       fill_value=fill)
-        with pytest.raises(ValueError):
-            out, mask = to_dense_batch(x, batch, batch_size=4, fill_value=fill)
-        with pytest.raises(ValueError):
+
+def test_to_dense_batch_disable_dynamic_shapes():
+    x = torch.Tensor([[1, 2], [3, 4], [5, 6], [7, 8], [9, 10], [11, 12]])
+    batch = torch.tensor([0, 0, 1, 2, 2, 2])
+
+    with set_experimental_mode(True, 'disable_dynamic_shapes'):
+        with pytest.raises(ValueError, match="'batch_size' needs to be set"):
+            out, mask = to_dense_batch(x, batch, max_num_nodes=6)
+        with pytest.raises(ValueError, match="'max_num_nodes' needs to be"):
+            out, mask = to_dense_batch(x, batch, batch_size=4)
+        with pytest.raises(ValueError, match="'batch_size' needs to be set"):
             out, mask = to_dense_batch(x)
 
-        out, mask = to_dense_batch(x, batch_size=1, max_num_nodes=6,
-                                   fill_value=fill)
+        out, mask = to_dense_batch(x, batch_size=1, max_num_nodes=6)
         assert out.size() == (1, 6, 2)
-        assert out.tolist() == [[[1, 2], [3, 4], [5, 6], [7, 8], [9, 10],
-                                 [11, 12]]]
-        assert mask.tolist() == [[1, 1, 1, 1, 1, 1]]
+        assert mask.size() == (1, 6)
 
-        out, mask = to_dense_batch(x, batch, batch_size=3, max_num_nodes=10,
-                                   fill_value=fill)
+        out, mask = to_dense_batch(x, batch, batch_size=3, max_num_nodes=10)
         assert out.size() == (3, 10, 2)
-        assert torch.equal(
-            out[0, :3], torch.Tensor([[1.0, 2.0], [3.0, 4.0], [item, item]]))
-        assert torch.equal(
-            out[1, :3], torch.Tensor([[5.0, 6.0], [item, item], [item, item]]))
-        assert torch.equal(
-            out[2, :3], torch.Tensor([[7.0, 8.0], [9.0, 10.0], [11.0, 12.0]]))
-        assert mask.tolist() == [[1, 1, 0, 0, 0, 0, 0, 0, 0, 0],
-                                 [1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                                 [1, 1, 1, 0, 0, 0, 0, 0, 0, 0]]
+        assert mask.size() == (3, 10)
 
 
 @onlyFullTest